requirements.txt
python-dotenv==1.0.0
pytest==7.4.3
requests==2.31.0
aiohttp==3.9.1
loguru==0.7.2
pydantic==2.5.2
typing-extensions==4.8.0
python-jose==3.3.0
cryptography==41.0.5
fastapi==0.104.1
uvicorn==0.24.0 